src/test/java/alf/exercises/leetcode/LongestPalindromicSubstringTest.java
package alf.exercises.leetcode;

import org.junit.Test;

import static org.junit.Assert.*;

public class LongestPalindromicSubstringTest {

    LongestPalindromicSubstring ex = new LongestPalindromicSubstring();

    @Test
    public void longestPalindrome() {
        assertEquals("bab", ex.longestPalindrome("babad"));
        assertEquals("bb", ex.longestPalindrome("cbbd"));
        assertEquals("bb", ex.longestPalindrome("abb"));
        assertEquals("xrcrx", ex.longestPalindrome("cyyoacmjwjubfkzrrbvquqkwhsxvmytmjvbborrtoiyotobzjmohpadfrvmxuagbdczsjuekjrmcwyaovpiogspbslcppxojgbfxhtsxmecgqjfuvahzpgprscjwwutwoiksegfreortttdotgxbfkisyakejihfjnrdngkwjxeituomuhmeiesctywhryqtjimwjadhhymydlsmcpycfdzrjhstxddvoqprrjufvihjcsoseltpyuaywgiocfodtylluuikkqkbrdxgjhrqiselmwnpdzdmpsvbfimnoulayqgdiavdgeiilayrafxlgxxtoqskmtixhbyjikfmsmxwribfzeffccczwdwukubopsoxliagenzwkbiveiajfirzvngverrbcwqmryvckvhpiioccmaqoxgmbwenyeyhzhliusupmrgmrcvwmdnniipvztmtklihobbekkgeopgwipihadswbqhzyxqsdgekazdtnamwzbitwfwezhhqznipalmomanbyezapgpxtjhudlcsfqondoiojkqadacnhcgwkhaxmttfebqelkjfigglxjfqegxpcawhpihrxydprdgavxjygfhgpcylpvsfcizkfbqzdnmxdgsjcekvrhesykldgptbeasktkasyuevtxrcrxmiylrlclocldmiwhuizhuaiophykxskufgjbmcmzpogpmyerzovzhqusxzrjcwgsdpcienkizutedcwrmowwolekockvyukyvmeidhjvbkoortjbemevrsquwnjoaikhbkycvvcscyamffbjyvkqkyeavtlkxyrrnsmqohyyqxzgtjdavgwpsgpjhqzttukynonbnnkuqfxgaatpilrrxhcqhfyyextrvqzktcrtrsbimuokxqtsbfkrgoiznhiysfhzspkpvrhtewthpbafmzgchqpgfsuiddjkhnwchpleibavgmuivfiorpteflholmnxdwewj"));
    }

    @Test
    public void isPalindrome() {
        assertTrue(ex.isPalindrome("bab"));
        assertTrue(ex.isPalindrome("bb"));
        assertTrue(ex.isPalindrome("cbbc"));
        assertTrue(ex.isPalindrome("badab"));

        assertFalse(ex.isPalindrome("bad"));
        assertFalse(ex.isPalindrome("babad"));
        assertFalse(ex.isPalindrome("cbbd"));
    }
}